requirements.txt
claude-agent-sdk>=0.1.58
openai>=1.0.0
watchfiles>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
requests>=2.33.1
psycopg2-binary>=2.9.0
PyJWT>=2.0.0